// FEBC75/src/types/CongViec.type.ts
export type CongViec = {
  id?:                    number;
  tenCongViec?:           string;
  danhGia?:               number;
  giaTien?:               number;
  nguoiTao?:              number;
  hinhAnh?:               string;
  moTa?:                  string;
  maChiTietLoaiCongViec?: number;
  moTaNgan?:              string;
  saoCongViec?:           number;
}